Move initialize function for hipparcos and celes info
