+++ src/environment/global/initialize_global_environment.cpp
@@ -10,76 +10,4 @@
 #include <environment/global/simulation_time.hpp>
 #include <library/initialize/initialize_file_access.hpp>
 
-HipparcosCatalogue* InitHipparcosCatalogue(std::string file_name) {
-  IniAccess ini_file(file_name);
-  const char* section = "HIPPARCOS_CATALOGUE";
 
-  std::string catalogue_path = ini_file.ReadString(section, "catalogue_file_path");
-  double max_magnitude = ini_file.ReadDouble(section, "max_magnitude");
-
-  HipparcosCatalogue* hipparcos_catalogue_;
-  hipparcos_catalogue_ = new HipparcosCatalogue(max_magnitude, catalogue_path);
-  hipparcos_catalogue_->IsCalcEnabled = ini_file.ReadEnable(section, CALC_LABEL);
-  hipparcos_catalogue_->is_log_enabled_ = ini_file.ReadEnable(section, LOG_LABEL);
-  hipparcos_catalogue_->ReadContents(catalogue_path, ',');
-
-  return hipparcos_catalogue_;
-}
-
-CelestialInformation* InitCelestialInformation(std::string file_name) {
-  IniAccess ini_file(file_name);
-  const char* section = "CELESTIAL_INFORMATION";
-  const char* furnsh_section = "CSPICE_KERNELS";
-
-  // Read SPICE setting
-  std::string inertial_frame = ini_file.ReadString(section, "inertial_frame");
-  std::string aber_cor = ini_file.ReadString(section, "aberration_correction");
-  std::string center_obj = ini_file.ReadString(section, "center_object");
-
-  // SPICE Furnsh
-  std::vector<std::string> keywords = {"tls", "tpc1", "tpc2", "tpc3", "bsp"};
-  for (size_t i = 0; i < keywords.size(); i++) {
-    std::string fname = ini_file.ReadString(furnsh_section, keywords[i].c_str());
-    furnsh_c(fname.c_str());
-  }
-
-  // Initialize celestial body list
-  const int num_of_selected_body = ini_file.ReadInt(section, "number_of_selected_body");
-  int* selected_body = new int[num_of_selected_body];
-  for (int i = 0; i < num_of_selected_body; i++) {
-    // Convert body name to SPICE ID
-    std::string selected_body_i = "selected_body_name(" + std::to_string(i) + ")";
-    char selected_body_temp[30];
-    ini_file.ReadChar(section, selected_body_i.c_str(), 30, selected_body_temp);
-    SpiceInt planet_id;
-    SpiceBoolean found;
-    bodn2c_c(selected_body_temp, (SpiceInt*)&planet_id, (SpiceBoolean*)&found);
-
-    // If the object specified in the ini file is not found, exit the program.
-    assert(found == SPICETRUE);
-
-    selected_body[i] = planet_id;
-  }
-
-  // Read Rotation setting
-  RotationMode rotation_mode;
-  std::string rotation_mode_temp = ini_file.ReadString(section, "rotation_mode");
-  if (rotation_mode_temp == "Idle") {
-    rotation_mode = RotationMode::kIdle;
-  } else if (rotation_mode_temp == "Simple") {
-    rotation_mode = RotationMode::kSimple;
-  } else if (rotation_mode_temp == "Full") {
-    rotation_mode = RotationMode::kFull;
-  } else  // if rotation_mode is neither Idle, Simple, nor Full, set rotation_mode to Idle
-  {
-    rotation_mode = RotationMode::kIdle;
-  }
-
-  CelestialInformation* celestial_info;
-  celestial_info = new CelestialInformation(inertial_frame, aber_cor, center_obj, rotation_mode, num_of_selected_body, selected_body);
-
-  // log setting
-  celestial_info->is_log_enabled_ = ini_file.ReadEnable(section, LOG_LABEL);
-
-  return celestial_info;
-}
