+++ src/environment/global/initialize_global_environment.hpp
@@ -10,18 +10,5 @@
 #include <environment/global/hipparcos_catalogue.hpp>
 #include <environment/global/simulation_time.hpp>
 
-/**
- *@fn InitHipCatalogue
- *@brief Initialize function for HipparcosCatalogue class
- *@param [in] file_name: Path to the initialize function
- */
-HipparcosCatalogue* InitHipparcosCatalogue(std::string file_name);
-
-/**
- *@fn InitCelesInfo
- *@brief Initialize function for CelestialInformation class
- *@param [in] file_name: Path to the initialize function
- */
-CelestialInformation* InitCelestialInformation(std::string file_name);
 
 #endif  // S2E_ENVIRONMENT_GLOBAL_INITIALIZE_GLOBAL_ENVIRONMENT_HPP_
